+++ VideoAPI/Video.API/Controllers/ConferenceController.cs
@@ -36,11 +36,13 @@ namespace Video.API.Controllers
         private readonly ServicesConfiguration _servicesConfiguration;
         private readonly ILogger<ConferenceController> _logger;
         private readonly IAudioPlatformService _audioPlatformService;
+        private readonly IStorageService _storageService;
 
 
         public ConferenceController(IQueryHandler queryHandler, ICommandHandler commandHandler,
             IVideoPlatformService videoPlatformService, IOptions<ServicesConfiguration> servicesConfiguration,
-            ILogger<ConferenceController> logger, IAudioPlatformService audioPlatformService)
+            ILogger<ConferenceController> logger, IAudioPlatformService audioPlatformService,
+            IStorageService storageService)
         {
             _queryHandler = queryHandler;
             _commandHandler = commandHandler;
@@ -48,6 +50,7 @@ namespace Video.API.Controllers
             _servicesConfiguration = servicesConfiguration.Value;
             _logger = logger;
             _audioPlatformService = audioPlatformService;
+            _storageService = storageService;
         }
 
         /// <summary>
@@ -57,12 +60,12 @@ namespace Video.API.Controllers
         /// <returns>Details of the new conference</returns>
         [HttpPost]
         [SwaggerOperation(OperationId = "BookNewConference")]
-        [ProducesResponseType(typeof(ConferenceDetailsResponse), (int) HttpStatusCode.Created)]
+        [ProducesResponseType(typeof(ConferenceDetailsResponse), (int)HttpStatusCode.Created)]
         [ProducesResponseType((int) HttpStatusCode.BadRequest)]
         public async Task<IActionResult> BookNewConferenceAsync(BookNewConferenceRequest request)
         {
             _logger.LogDebug("BookNewConference");
-            
+
             foreach (var participant in request.Participants)
             {
                 participant.Username = participant.Username.ToLower().Trim();
@@ -81,21 +84,21 @@ namespace Video.API.Controllers
             {
                 _logger.LogWarning($"Error creating audio recording for caseNumber: {request.CaseNumber} and hearingId: {request.HearingRefId}");
             }
-            
+
             var conferenceId = await CreateConferenceAsync(request, createAudioRecordingResponse.IngestUrl);
             _logger.LogDebug("Conference Created");
-            
+
             await BookKinlyMeetingRoomAsync(conferenceId, request.AudioRecordingRequired, createAudioRecordingResponse.IngestUrl);
             _logger.LogDebug("Kinly Room Booked");
-            
+
             var getConferenceByIdQuery = new GetConferenceByIdQuery(conferenceId);
             var queriedConference = await _queryHandler.Handle<GetConferenceByIdQuery, Conference>(getConferenceByIdQuery);
 
             var response = ConferenceToDetailsResponseMapper.MapConferenceToResponse(queriedConference, _servicesConfiguration.PexipSelfTestNode);
-            
+
             _logger.LogInformation($"Created conference {response.Id} for hearing {request.HearingRefId}");
-            
-            return CreatedAtAction(nameof(GetConferenceDetailsByIdAsync), new {conferenceId = response.Id}, response);
+
+            return CreatedAtAction(nameof(GetConferenceDetailsByIdAsync), new { conferenceId = response.Id }, response);
         }
 
         /// <summary>
@@ -120,9 +123,9 @@ namespace Video.API.Controllers
 
                 return NotFound();
             }
-            
+
             await _videoPlatformService.UpdateVirtualCourtRoomAsync(conference.Id, request.AudioRecordingRequired);
-            
+
             try
             {
                 var command = new UpdateConferenceDetailsCommand(request.HearingRefId, request.CaseNumber,
@@ -130,7 +133,7 @@ namespace Video.API.Controllers
                     request.HearingVenueName, request.AudioRecordingRequired);
 
                 await _commandHandler.Handle(command);
-                
+
                 return Ok();
             }
             catch (ConferenceNotFoundException)
@@ -152,17 +155,17 @@ namespace Video.API.Controllers
         public async Task<IActionResult> GetConferenceDetailsByIdAsync(Guid conferenceId)
         {
             _logger.LogDebug($"GetConferenceDetailsById {conferenceId}");
-            
+
             var getConferenceByIdQuery = new GetConferenceByIdQuery(conferenceId);
             var queriedConference = await _queryHandler.Handle<GetConferenceByIdQuery, Conference>(getConferenceByIdQuery);
 
             if (queriedConference == null)
             {
                 _logger.LogWarning($"Unable to find conference {conferenceId}");
-                
+
                 return NotFound();
             }
-            
+
             var response = ConferenceToDetailsResponseMapper.MapConferenceToResponse(queriedConference, _servicesConfiguration.PexipSelfTestNode);
             return Ok(response);
         }
@@ -185,15 +188,15 @@ namespace Video.API.Controllers
             {
                 await _commandHandler.Handle(removeConferenceCommand);
                 await SafelyRemoveCourtRoomAsync(conferenceId);
-                
+
                 _logger.LogInformation($"Successfully removed conference {conferenceId}");
-                
+
                 return NoContent();
             }
             catch (ConferenceNotFoundException ex)
             {
                 _logger.LogError(ex, $"Unable to find conference {conferenceId}");
-                
+
                 return NotFound();
             }
         }
@@ -209,16 +212,16 @@ namespace Video.API.Controllers
             [FromQuery] ConferenceForAdminRequest request)
         {
             _logger.LogDebug("GetConferencesTodayForAdmin");
-            
+
             var query = new GetConferencesTodayForAdminQuery
             {
                 UserNames = request.UserNames
             };
-            
+
             var conferences = await _queryHandler.Handle<GetConferencesTodayForAdminQuery, List<Conference>>(query);
 
             var response = conferences.Select(ConferenceForAdminResponseMapper.MapConferenceToSummaryResponse);
-            
+
             return Ok(response);
         }
 
@@ -234,20 +237,20 @@ namespace Video.API.Controllers
         public async Task<IActionResult> GetConferencesTodayForJudgeByUsernameAsync([FromQuery] string username)
         {
             _logger.LogDebug($"GetConferencesTodayForJudgeByUsername {username}");
-            
+
             if (!username.IsValidEmail())
             {
                 ModelState.AddModelError(nameof(username), $"Please provide a valid {nameof(username)}");
-                
+
                 _logger.LogWarning($"Invalid username {username}");
-                
+
                 return BadRequest(ModelState);
             }
 
             var query = new GetConferencesForTodayByJudgeQuery(username.ToLower().Trim());
             var conferences = await _queryHandler.Handle<GetConferencesForTodayByJudgeQuery, List<Conference>>(query);
             var response = conferences.Select(ConferenceForJudgeResponseMapper.MapConferenceSummaryToModel);
-            
+
             return Ok(response);
         }
 
@@ -263,13 +266,13 @@ namespace Video.API.Controllers
         public async Task<IActionResult> GetConferencesTodayForIndividualByUsernameAsync([FromQuery] string username)
         {
             _logger.LogDebug($"GetConferencesTodayForIndividualByUsername {username}");
-            
+
             if (!username.IsValidEmail())
             {
                 ModelState.AddModelError(nameof(username), $"Please provide a valid {nameof(username)}");
-                
+
                 _logger.LogWarning($"Invalid username {username}");
-                
+
                 return BadRequest(ModelState);
             }
 
@@ -292,38 +295,38 @@ namespace Video.API.Controllers
         public async Task<IActionResult> GetConferenceByHearingRefIdAsync(Guid hearingRefId)
         {
             _logger.LogDebug($"GetConferenceByHearingRefId {hearingRefId}");
-            
+
             var query = new GetConferenceByHearingRefIdQuery(hearingRefId);
             var conference = await _queryHandler.Handle<GetConferenceByHearingRefIdQuery, Conference>(query);
 
             if (conference == null)
             {
                 _logger.LogWarning($"Unable to find conference with hearing id {hearingRefId}");
-                
+
                 return NotFound();
             }
-            
+
             var response = ConferenceToDetailsResponseMapper.MapConferenceToResponse(conference, _servicesConfiguration.PexipSelfTestNode);
-            
+
             return Ok(response);
         }
-        
+
         /// <summary>
         /// Get list of expired conferences 
         /// </summary>
         /// <returns>Conference summary details</returns>
-        
+
         [HttpGet("expired")]
         [SwaggerOperation(OperationId = "GetExpiredOpenConferences")]
         [ProducesResponseType(typeof(List<ExpiredConferencesResponse>), (int) HttpStatusCode.OK)]
         public async Task<IActionResult> GetExpiredOpenConferencesAsync()
         {
             _logger.LogDebug("GetExpiredOpenConferences");
-            
+
             var query = new GetExpiredUnclosedConferencesQuery();
             var conferences = await _queryHandler.Handle<GetExpiredUnclosedConferencesQuery, List<Conference>>(query);
             var response = conferences.Select(ConferenceToExpiredConferenceMapper.MapConferenceToExpiredResponse);
-            
+
             return Ok(response);
         }
 
@@ -345,17 +348,17 @@ namespace Video.API.Controllers
                 await _commandHandler.Handle(command);
                 await SafelyRemoveCourtRoomAsync(conferenceId);
                 await DeleteAudioRecordingApplication(conferenceId);
-                
+
                 return NoContent();
             }
             catch (ConferenceNotFoundException ex)
             {
                 _logger.LogError(ex, $"Unable to find conference {conferenceId}");
-                
+
                 return NotFound();
             }
         }
-        
+
         /// <summary>
         /// Get today's conferences where judges are in hearings
         /// </summary>
@@ -370,7 +373,7 @@ namespace Video.API.Controllers
             var conferences = await _queryHandler.Handle<GetJudgesInHearingsTodayQuery, List<Conference>>(new GetJudgesInHearingsTodayQuery());
 
             var response = conferences.SelectMany(ConferenceForJudgeResponseMapper.MapConferenceSummaryToJudgeInHearingResponse);
-            
+
             return Ok(response);
         }
 
@@ -390,8 +393,23 @@ namespace Video.API.Controllers
 
             if (queriedConference != null && queriedConference.AudioRecordingRequired)
             {
-                await _audioPlatformService.DeleteAudioApplicationAsync(queriedConference.HearingRefId);
+                if (await CheckAudioRecordingFile(queriedConference.HearingRefId))
+                {
+                    await _audioPlatformService.DeleteAudioApplicationAsync(queriedConference.HearingRefId);
+                }
+                else
+                {
+                    var msg = $"Audio recording file not found for hearing: {queriedConference.HearingRefId}";
+                    var ex = new AudioPlatformException(msg, HttpStatusCode.NotFound);
+                    _logger.LogError(ex, msg);
+                }
             }
+        }
+
+        private async Task<bool> CheckAudioRecordingFile(Guid hearingId)
+        {
+            var filePath = $"{hearingId}.mp4";
+            return await _storageService.FileExistsAsync(filePath);
         }
 
         private async Task BookKinlyMeetingRoomAsync(Guid conferenceId, bool audioRecordingRequired, string ingestUrl)
@@ -399,22 +417,22 @@ namespace Video.API.Controllers
             MeetingRoom meetingRoom;
             try
             {
-                meetingRoom = await _videoPlatformService.BookVirtualCourtroomAsync(conferenceId, audioRecordingRequired, ingestUrl);            
+                meetingRoom = await _videoPlatformService.BookVirtualCourtroomAsync(conferenceId, audioRecordingRequired, ingestUrl);
             }
             catch (DoubleBookingException ex)
             {
                 _logger.LogError(ex, $"Kinly Room already booked for conference {conferenceId}");
-                
+
                 meetingRoom = await _videoPlatformService.GetVirtualCourtRoomAsync(conferenceId);
             }
 
             if (meetingRoom == null) return;
-            
+
             var command = new UpdateMeetingRoomCommand
             (
                 conferenceId, meetingRoom.AdminUri, meetingRoom.JudgeUri, meetingRoom.ParticipantUri, meetingRoom.PexipNode
             );
-            
+
             await _commandHandler.Handle(command);
         }
 
@@ -424,7 +442,7 @@ namespace Video.API.Controllers
                 new CheckConferenceOpenQuery(request.ScheduledDateTime, request.CaseNumber, request.CaseName));
 
             if (existingConference != null) return existingConference.Id;
-            
+
             var participants = request.Participants.Select(x =>
                     new Participant(x.ParticipantRefId, x.Name, x.FirstName, x.LastName, x.DisplayName, x.Username, x.UserRole,
                         x.CaseTypeGroup)
@@ -432,15 +450,15 @@ namespace Video.API.Controllers
                         Representee = x.Representee
                     })
                 .ToList();
-            
+
             var createConferenceCommand = new CreateConferenceCommand
             (
-                request.HearingRefId, request.CaseType, request.ScheduledDateTime, request.CaseNumber, 
+                request.HearingRefId, request.CaseType, request.ScheduledDateTime, request.CaseNumber,
                 request.CaseName, request.ScheduledDuration, participants, request.HearingVenueName, request.AudioRecordingRequired, ingestUrl
             );
-            
+
             await _commandHandler.Handle(createConferenceCommand);
-            
+
             return createConferenceCommand.NewConferenceId;
         }
     }
